Average price & Delete button
In the Home screen users are now able to view average prices of menu items broken down into different courses/category. And in the Add Menu Screen users are able to delete menu items from the list that are still in the Home Screen. You can view the ReadMe document to gain more information.